cleaned up channel
